fix: context warnings
