+++ src/context/auth-data.ts
@@ -1,0 +1,6 @@
+import { defaultUsername } from '../shared/constants';
+
+export const defaultAuth = {
+  isAuthenticated: false,
+  username: defaultUsername,
+};
