+++ src/context/types.ts
@@ -1,0 +1,10 @@
+import { Dispatch, SetStateAction } from 'react';
+
+export type AuthState = {
+  isAuthenticated: boolean;
+  username: string;
+};
+
+export type AuthContextProps = AuthState & {
+  setAuth: Dispatch<SetStateAction<AuthState>>;
+};
